django_server/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
defusedxml==0.7.1
Django==4.0.1
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-api-key==2.1.0
djangorestframework-camel-case==1.3.0
djangorestframework-simplejwt==5.0.0
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
